requirements.txt
transformers==4.38.1
pandas==2.2.0
python-dotenv
torch
